Enable trackbar seeking
Changed update frequency of trackbar to reduce crashing of that control. Also, enabled the trackbar and added a listener for the WM_HSCROLL message so that it will enable me to know when the user is scrolling the bar. Scrolling the bar will now update the video's position to match.